+++ packages/survey_kit/example/android/app/src/main/kotlin/com/quickbirdstudios/example/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.quickbirdstudios.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity()