+++ packages/survey_kit/lib/src/util/extension.dart
@@ -1,7 +1,13 @@
 extension JoinedIterable<T> on Iterable<T> {
-  Iterable<T> withSeparator(T separator) {
-    return isEmpty
-        ? this
-        : (expand((element) => [element, separator]).toList());
+  Iterable<T> withSeparator(T separator) sync* {
+    final iterator = this.iterator;
+    if (!iterator.moveNext()) {
+      return;
+    }
+    yield iterator.current;
+    while (iterator.moveNext()) {
+      yield separator;
+      yield iterator.current;
+    }
   }
 }