+++ packages/survey_kit/lib/src/view/widget/answer/answer_view.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart' hide Step;
-import 'package:survey_kit/src/configuration/survey_kit_registry.dart';
 import 'package:survey_kit/src/view/step_view.dart';
 import 'package:survey_kit/survey_kit.dart';